+++ src/main/java/com/endava/pocu/gamesApp/services/GameService.java
@@ -1,5 +1,8 @@
 package com.endava.pocu.gamesApp.services;
 
+import com.endava.pocu.gamesApp.exceptions.InvalidRequestException;
+import com.endava.pocu.gamesApp.exceptions.RecordNotFoundException;
+import com.endava.pocu.gamesApp.models.Achievement;
 import com.endava.pocu.gamesApp.models.Game;
 import com.endava.pocu.gamesApp.repositories.GameRepository;
 import org.springframework.beans.BeanUtils;
@@ -17,6 +20,8 @@ public class GameService {
     private GameRepository gameRepository;
 
     public Game saveGame(final Game game){
+        if (game == null)
+            throw new InvalidRequestException("Game must not be null");
         return gameRepository.saveAndFlush(game);
     }
 
@@ -27,6 +32,8 @@ public class GameService {
     }
 
     public Optional<List<Game>> findByTagTitle(String tagTitle){
+        if (tagTitle == null)
+            throw new InvalidRequestException("Tag title must not be null");
         return gameRepository.findGamesByTagTitle(tagTitle);
     }
 
@@ -35,15 +42,27 @@ public class GameService {
         if (game.isPresent())
             return game.get();
         else
-            throw new RuntimeException("Game with id "+id+" does not exist");
+            throw new RecordNotFoundException("Game with id "+id+" does not exist");
     }
 
     public Game updateGame(Long id, Game game){
-            Game existingGame = gameRepository.getReferenceById(id);
-            BeanUtils.copyProperties(game, existingGame, "gameId");
-            return gameRepository.saveAndFlush(existingGame);
+
+        if (id == null || game == null)
+            throw new InvalidRequestException("Game and id must not be null");
+        Optional<Game> optionalExistingGame = gameRepository.findById(id);
+        if (optionalExistingGame.isEmpty())
+            throw new RecordNotFoundException("Game with id "+id+" does not exist");
+
+        Game existingGame = optionalExistingGame.get();;
+        BeanUtils.copyProperties(game, existingGame, "gameId");
+        return gameRepository.saveAndFlush(existingGame);
     }
     public void deleteGame(final Long id){
+        if (id == null)
+            throw new InvalidRequestException("Id must not be null");
+        Optional<Game> optionalExistingGame = gameRepository.findById(id);
+        if (optionalExistingGame.isEmpty())
+            throw new RecordNotFoundException("Game with id "+id+" does not exist");
         gameRepository.deleteById(id);
     }
 }
